Replace eval-core and current-primitive-locations with rnrs eval
The current-primitive-locations and eval-core procedures work together
to make it possible for psyntax to run R6RS code while expanding
macros. The expression rewriter in (r6lint psyntax internal) uses
current-primitive-locations to find the locations of primitives. Now
eval-core is just eval from (rnrs eval) together with an environment
that contains most primitives.

Primitives that can write to the file system are not available right
now. It's one thing for macros to access the file system when a
program is being compiled, but it's quite another to have it happen
when a file is just opened in an editor. For similar reasons the
macros don't have access to eval, although they really should have it.
Hopefully there are not that many macros that use eval or write to the
file system, so this should be fine for now.

Additionally instead of using top-level-value as a substitute for
symbol-value, now the symbol property 'value is used. This probably
fixes a memory leak.