Add an option to psyntax for using r6rs eval to run core forms
This adds an option that puts global variables in core forms into a
hashtable. These are used by psyntax when running macro expanders.
Normally code runs in a top-level environment, but that is not
portable. This change should make it possible to run psyntax on top of
basically any R6RS Scheme.

This commit also includes some fixes to the (rnrs) library definitions
that came with psyntax.