Add a fault tolerant mode to the reader
The reader has a new tolerant mode that is enabled by the linter. In
case of lexical violations a continuable warning is raised and the
returned datum is an approximation of what was in the source. In most
cases the datum should still be possible to macro expand. The end
result is a linter that shows more than just the first lexical
violation.