Keep source information from the expander, warn about unused vars
The output language of the expander is now (optionally) based on
records. These records maintain source information if the expander has
kept the source.

A new analyser has been added that checks for unused variables. Some
macros generate false positives for now, but apart from that the
linter shows what variables are unused. The linter doesn't print
anything for variables starting with an underscore (_).