Portability fixes to support more than Chez
This fixes a few issues that were making the expander unportable. The
expanded code no longer needs the void primitive. Instead explicitly
one-armed conditionals are built. Gensyms are supported using a trick
to use the host's own gensym (not sure how portable it is in the end,
though).

If the host has a good gensym that remembers the original name for the
symbol, then that should really be used instead.